+++ CHANGES.rst
@@ -1,12 +1,13 @@
 CHANGES
 =======
 
-3.x
----
+3.2 2018-11-18
+--------------
 
 * Issue 17: Feature request: simple query output format to ease goobook use with notmuch
 * dquery: Don't print header if there is no groups.
 * Issue 69: Added note about regexps to man page.
+* Issue 79: Fixed parsing of birthdays without date (fix is to ignore them)
 
 3.1 2018-10-28
 --------------